Restructure the ig folder and renamed it to dictionaries
