+++ dictionaries/utils/writeToFiles.js
